Ported most of ~nicoleaoki's pages to Markdown
This commit spearheads the ongoing transition from HTML-written pages to Markdown-written pages in ~nicoleaoki, soon all future pages written for this site will be made using Markdown.

Index is not getting ported, About is being rewritten so that's why I didn't copy and paste it, and Contributing is getting ported later as that's a pain in the ass to do and I'm going to add more things to it anyways.

Also yeeted the newlaptop article and made it a .zip incase you're really interested in reading it for some reason.